Add feature for sending `feature_report`
This change tends to add functional for sending feature report from device to the host.

There are a lot of HID devices in which report descriptors are specified feature reports sections. Currently it is not available to send feature_report from device to host using USB_GADGET  interface.  The HOST side may require to get feature report from device for enumerating it properly and setting control options to the device depending from HID_REPORT descriptor definition.